use a specific time, not system time
When we call Time.new without any arguments, the time generated will be
the time from your computer system at the point you called Time.new.

We can simplify our test by creating a Time object with arguments which
specify the year, month, day, hour and minute that we want. This way we
don't need to use the format_time method in our test and our test is
cleaner as it is only testing one specific method (the message method on
Text) and we can specify the time in the test without using string
interpolation.

Simpler. Better.