Use a stub instead of a Time object
In this variation, we use a stub instead of an actual Time object.
Otherwise, this is the same as the code in commit f59c67.

We use Rspec's Mocks library to create a double of Time. In this case,
it's an 'instance double' of Time, which means that it is not actually a
Time object, but Rspec will verify that any methods we stub on this
double will be verified to exist on an actual Time object. This way we
can provide canned responses to an object which represents Time. In our
case, we provide a canned response to Time's 'hour' method and also the
'min' method, using the following Rsepc syntax:

allow(time).to receive(:hour).and_return(15)

This means that every time the 'hour' method is called, the response
from the stub will be the number 15. We can guarantee then, as our code
in Text will add 1, that the message hour to be used in our text message
will be 16.

This is conceptually a little different to using a real Time object
which has a fixed time. By using the real Time object, we also check
that a real instance of Time actually works with a real instance of
Text. For this reason the test could be considered to be a form of
'integration' test (ie. you are testing the integration of Text and
Time). If we  want to truly isolate Text from Time in our test, we can
use a double (in this case, specifically, a stub). There are perfectly
good reasons for using either approach - you could read up on London
style vs. Chicago style TDD for a big of background on each.