refactor test to use 'let' syntax
This is actually no different to the previous commit, other than the
fact that if we had multiple tests in this file we would be able to use
the Time double in every test without having to declare it in each test.
Functionally speaking though it's the same as we only have one test, we
create an instance_double of Time and assign it to a variable called
time.