use a hand-crafted stub
Here we use a hand crafted stub. This isn't necessarily advisable in
Ruby, but it's absolutely possible, due to duck typing.

At the end of the day, Text will accept any object that has an 'hour'
method and a 'min' method, so long as those methods return a value which
is compatible with the code in the Text class. Instead of using Rspec to
generate a stub for Time, I have added my own TimeStub class, which has
the required 'hour' and 'min' methods. You can see that they simply
return 15 and 37 respectively, so this behaves in exactly the same way
as the Rspec generated stub.

There is a difference, as Rspec will also verify that the methods
stubbed are actually present on Time, and our hand crafted stub doesn't
do that, but otherwise it's really the same thing as using Rspec to do
it.

It is also (hopefully) a good example of duck typing. As far as Ruby is
concerned, if it walks like a duck and talks like a duck, it IS a duck!