add script/entrypoint to run the program
So far we have a collection of classes, but no way to run them, other
than by 'require'-ing them all in IRB and manually making them interact
with each other.

It's important to note that there is absolutely no requirement for
any other way to execute the code, according to the requirements
document. In fact, it even suggests that the classes can be instantiated
and made to interact in IRB. It is dangerous to add requirements to a
project that haven't been asked for. In a tech test you may even fail
the test for this reason. However, for a real world program, it is quite
unsatisfactory to have to use some classes directly via IRB.

This commit is designed to show how a collection of classes can be
assembled into a working program that can be used on the command line.
It is done in a fairly simple way, it could definitely be a bit nicer,
but it demonstrates the general idea.

Using the program from the command line hopefully also demonstrates the idea that using 'puts' in your
classes is usually not ideal, particularly for a command line
application. Ideally your user interface code should be collected
together in one place so you don't have to go scouring your classes to
find where that rogue 'puts' is located. Classes that are not concerned
with your user interface should usually just take
inputs and generate outputs.

Using the program as a real user would is also a good way of verifying
whether it really works in a way that you want it to, you may realise
while using it that there was a required feature that you missed somehow
during development. It happens.

The other thing worth noting is that the script (or entrypoint, there
are a series of terms that you might want to use to describe this file)
is in the 'bin' directory. This is a convention for executable files, it
would work just the same in any other directory (as long as the
require_relative paths were updated accordingly), there's no magic
significance to the bin directory other than convention.

As the file is executable and has a hashbang at the top of the file,
executable permissions have been added to it (chmod +x <filename>). It is then run
from the root level of the project by entering:

./bin/takeaway.rb