Order can take a menu, fix failing tests
In order to properly stub the Menu when testing the Order class, we need
to pass a Menu in to the Order. This is essential for testing. Calling
Menu.new in the constructor of Order is going to make testing near
enough impossible.

Once it's passed in, it's easy to create a double of Menu and then use
the Rspec to allow the double to receive calls to menu_list and return
any kind of canned response, which can then be tested against.