Don't use puts!
It's bad style to use puts in your classes, generally speaking. If you
can, it's better style to write methods that take data as inputs and
return data as outputs. In this case, it's arguable that there's no data
to return as an output, so I've added an attr_reader on @item so that I
can query the items that have been ordered. This makes it easy to write
a test that tests the adding of items without having to rely on 'puts'.

It's OK for the user of the program to have a message 'puts'-ed onto
their console, but this should be done wherever the other user interface
code is.