Pass in dishes to Menu
In it's previous state, Menu had a hardcoded list of dishes. This is
problematic as it means that testing any methods on Menu is difficult.
As we keep finding out, in order to test things we really need to be
able to pass their dependencies in. It also makes our code generally
more flexible.

The previous test was also a little bit vaccuous, since we're just testing an
instance method on Menu. Personally I would argue that it's unnecessary,
but I know people in the Ruby world like to do things a bit differently,
and I defend their right to do so. In that spirit I've left in a test on
the instance method. It's much nicer now though, since we can just pass
in any old list of dishes and expect to get the same smaller one back.

I've also gotten a bit bold and decided that I want to get back the list
of dishes in a more human readable format. There was a printed_list
method that was unfinished so I've just finished a working version of
that. I think it could be nicer, but it works, so I'm just going to
leave it as it is for now.